ajout ft_printf dans libft + modif ft_strlen en type int

+++ srcs/ft_strdup.c
@@ -6,7 +6,7 @@
 /*   By: sofiahechaichi <sofiahechaichi@student.    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/10/13 18:06:45 by sohechai          #+#    #+#             */
-/*   Updated: 2021/02/02 11:01:02 by sofiahechai      ###   ########lyon.fr   */
+/*   Updated: 2021/02/02 11:19:04 by sofiahechai      ###   ########lyon.fr   */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,7 +19,7 @@ char	*ft_strdup(const char *src)
 	int		i;
 
 	i = 0;
-	length = ft_strlen(src);
+	length = (int)ft_strlen(src);
 	if (!(dest = malloc(sizeof(char) * (length) + 1)))
 		return (NULL);
 	while (src[i])
